Revert "Linux 3.10.32"
This reverts commit ba84ba5b64751afb7b37a27aaf65c16a0ee9ae95.

+++ drivers/hwmon/ntc_thermistor.c
@@ -145,7 +145,7 @@ struct ntc_data {
 static int ntc_adc_iio_read(struct ntc_thermistor_platform_data *pdata)
 {
 	struct iio_channel *channel = pdata->chan;
-	s64 result;
+	unsigned int result;
 	int val, ret;
 
 	ret = iio_read_channel_raw(channel, &val);
@@ -155,10 +155,10 @@ static int ntc_adc_iio_read(struct ntc_thermistor_platform_data *pdata)
 	}
 
 	/* unit: mV */
-	result = pdata->pullup_uv * (s64) val;
+	result = pdata->pullup_uv * val;
 	result >>= 12;
 
-	return (int)result;
+	return result;
 }
 
 static const struct of_device_id ntc_match[] = {
